[TASK] Extract functionality into service
Functionality is extracted into a service class for re-usability in external controllers.
